refactor: extract shared LocomotionDRProvider base class (#184)
Introduce LocomotionDRProvider in common/dr_provider.py with shared
validate, build_interval_randomization_plan, _sample_commands, and
template methods for build_reset_plan/build_reset_observation.

G1/Go1/Go2 providers now inherit from the base class and only override
robot-specific hooks (_get_qvel_limit, _build_extra_info_updates,
_compute_reset_obs), eliminating ~150 lines of duplicated DR code.
