+++ src/unilab/envs/locomotion/common/__init__.py
@@ -1,6 +1,7 @@
 from .base import ControlConfigBase, LocomotionBaseCfg, LocomotionBaseEnv, Sensor
 from .commands import Commands, sample_velocity_commands
 from .domain_rand import DomainRandConfig
+from .dr_provider import LocomotionDRProvider
 from .rewards import RewardContext
 
 __all__ = [
@@ -9,6 +10,7 @@ __all__ = [
     "DomainRandConfig",
     "LocomotionBaseCfg",
     "LocomotionBaseEnv",
+    "LocomotionDRProvider",
     "RewardContext",
     "Sensor",
     "sample_velocity_commands",
